+++ packages/plugin-infra/src/sns-adaptor/index.ts
@@ -2,7 +2,6 @@ export * from './CompositionContext.js'
 export * from './getTabContent.js'
 export * from './getSearchResultTabs.js'
 export * from './PostContext.js'
-export * from './SNSAdaptorContext.js'
 export * from './useAllPersonas.js'
 export * from './useCurrentPersonaInformation.js'
 export * from './useCurrentVisitingIdentity.js'
