+++ packages/plugin-infra/src/sns-adaptor/useAllPersonas.ts
@@ -1,6 +1,6 @@
 import { useSubscription } from 'use-subscription'
 import { EMPTY_ARRAY } from '@masknet/shared-base'
-import { useSNSAdaptorContext } from './SNSAdaptorContext.js'
+import { useSNSAdaptorContext } from '../dom/useSNSAdaptorContext.js'
 
 export function useAllPersonas() {
     const { allPersonas } = useSNSAdaptorContext()
