+++ packages/plugin-infra/src/sns-adaptor/useCurrentPersonaInformation.ts
@@ -1,6 +1,6 @@
-import { UNDEFINED } from '@masknet/shared-base'
 import { useSubscription } from 'use-subscription'
-import { useSNSAdaptorContext } from '../sns-adaptor/SNSAdaptorContext.js'
+import { UNDEFINED } from '@masknet/shared-base'
+import { useSNSAdaptorContext } from '../dom/useSNSAdaptorContext.js'
 
 export function useCurrentPersonaInformation() {
     const { currentPersona, allPersonas } = useSNSAdaptorContext()
