+++ packages/plugin-infra/src/sns-adaptor/useCurrentVisitingIdentity.ts
@@ -1,6 +1,6 @@
 import { useSubscription } from 'use-subscription'
 import { UNDEFINED } from '@masknet/shared-base'
-import { useSNSAdaptorContext } from './SNSAdaptorContext.js'
+import { useSNSAdaptorContext } from '../dom/useSNSAdaptorContext.js'
 
 export function useCurrentVisitingIdentity() {
     const { currentVisitingProfile } = useSNSAdaptorContext()
