+++ packages/plugin-infra/src/sns-adaptor/useSocialIdentity.ts
@@ -2,7 +2,7 @@ import { useEffect } from 'react'
 import { useAsyncRetry } from 'react-use'
 import type { SocialIdentity } from '@masknet/shared-base'
 import type { IdentityResolved } from '../types.js'
-import { useSNSAdaptorContext } from './SNSAdaptorContext.js'
+import { useSNSAdaptorContext } from '../dom/useSNSAdaptorContext.js'
 
 /**
  * Get the social identity of the given identity
