+++ packages/plugins/Claim/src/SNSAdaptor/index.tsx
@@ -1,11 +1,10 @@
+import { useState } from 'react'
+import { Trans } from 'react-i18next'
 import { Icons } from '@masknet/icons'
 import type { Plugin } from '@masknet/plugin-infra'
-import { SNSAdaptorContext } from '@masknet/plugin-infra/content-script'
 import { PluginID } from '@masknet/shared-base'
 import { useRemoteControlledDialog } from '@masknet/shared-base-ui'
 import { isValidAddress } from '@masknet/web3-shared-evm'
-import { useState } from 'react'
-import { Trans } from 'react-i18next'
 import { base } from '../base.js'
 import { PLUGIN_ID } from '../constants.js'
 import { PluginClaimMessage } from '../message.js'
@@ -42,12 +41,12 @@ const sns: Plugin.SNSAdaptor.Definition = {
         )
 
         return (
-            <SNSAdaptorContext.Provider value={context}>
+            <>
                 {claimOpen ? <ClaimDialog open onClose={closeClaimDialog} /> : null}
                 {successOpen ? (
                     <ClaimSuccessDialog open onClose={closeSuccessDialog} tokenAddress={tokenAddress} amount={amount} />
                 ) : null}
-            </SNSAdaptorContext.Provider>
+            </>
         )
     },
     ApplicationEntries: [
