--- packages/plugins/Gitcoin/src/settings.ts
@@ -1,3 +1,0 @@
-import type { Plugin } from '@masknet/plugin-infra/content-script'
-
-export const SharedContext: { value: Plugin.SNSAdaptor.SNSAdaptorContext | undefined } = { value: undefined }
